Add config for disabling login methods
introduces new env settings:
OAUTH_LOGIN_ENABLED=true
PW_LOGIN_ENABLED=true
