+++ config/services.php
@@ -36,17 +36,10 @@ return [
     ],
 
     // OAuth Providers
-    'github' => [
-        'client_id' => env('GITHUB_CLIENT_ID'),
-        'client_secret' => env('GITHUB_CLIENT_SECRET'),
-        'redirect' => env('APP_URL') . '/oauth/github/callback',
-    ],
-
     'schulcloud' => [
         'client_id' => env('SCHULCLOUD_CLIENT_ID'),
         'client_secret' => env('SCHULCLOUD_CLIENT_SECRET'),
         'hydra_url' => env('SCHULCLOUD_HYDRA_URL', 'https://schul-cloud.org/hydra'),
         'redirect' => env('APP_URL') . '/oauth/schulcloud/callback',
     ],
-
 ];
